Do not try unregister original keyboard layouts resource
The path to the default keyboard layouts does not exist on NixOS and
possibly other systems. However, registering and unregistering the
default keyboard layouts is not needed as the registered modified
layouts resource is accessed with higher priority than the original
layouts resource until it is unregistered.

Additionally, the resources were not being unregistered since a
reference to the original Gio.Resource object is needed to unregister
it. Creating a new resource by loading it from the same path does not
create the same resource reference. The unregistering appeared to work
because new resources were continuously registered above the old ones.